# ticket-master/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/Account": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "Account_Register",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "model",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateAccountVM"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Account"
        ],
        "operationId": "Account_Login",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "model",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginVM"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Account"
        ],
        "operationId": "Account_Uvrede",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "uvreda",
            "in": "query",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {}
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "CreateAccountVM": {
      "type": "object",
      "required": [
        "isOrganisation"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "isOrganisation": {
          "type": "boolean"
        },
        "user": {
          "$ref": "#/definitions/CreateAccountUserVM"
        },
        "organisation": {
          "$ref": "#/definitions/CreateAccountOrganisationVM"
        }
      }
    },
    "CreateAccountUserVM": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "birthdate": {
          "type": "string",
          "format": "date-time"
        },
        "gender": {
          "type": "string"
        }
      }
    },
    "CreateAccountOrganisationVM": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "incorporatedDate": {
          "type": "string",
          "format": "date-time"
        },
        "description": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "addressLine1": {
          "type": "string"
        },
        "addressLine2": {
          "type": "string"
        },
        "addressLine3": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "zipCode": {
          "type": "string"
        }
      }
    },
    "LoginVM": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
  }
}